fix: move TestRoomQuerySetWithoutMembers to its own file (one class per file rule)
